Op.8 No.2: proofreading
Measure 12, beat 2, right hand: The final note of this beat should be
transposed up an octave in the higher voice to coincide with the note in
the lower voice.

Measure 13, beat 1, right hand: The first note of the measure in the
lower voice should be F-sharp, coinciding with the higher voice.